introduce callbacks for jobs and API design revised
When running a job, developer can define callbacks for stdout and
end_of_job.
When the job is over files are transferred to gn_server
(transfer_files/2).
When the job start a directory in the local path is created,
computation starts in that directory, files are transferred and the
computation directory is then removed recursively.